fixed: move the chrom-storage to extension-shared and remove chrome from package code
Closes #8
