+++ packages/data-model/src/cache-data-access.ts
@@ -4,9 +4,8 @@ import { type CacheDataItem } from './data-cache-types';
 import { storage } from './storage';
 import { type StorageChange } from './storage/storage-types';
 
-
 // The listeners object is used to store a reference to the listener function that is created by _updateCaller.
-// This is necessary because chrome.storage.onChanged.removeListener requires the exact same function instance
+// This is necessary because storage.removeStorageListener requires the exact same function instance
 // that was passed to addListener. If we call _updateCaller again in removeCachedDataListener, it will create
 // a new function instance, and the listener will not be removed.
 const listeners: {
@@ -68,7 +67,7 @@ export const addCachedDataListener = (
   updateCallback: (key: string, data: unknown) => void
 ) => {
   listeners[key] = _updateCaller(key, updateCallback);
-  chrome.storage.onChanged.addListener(listeners[key]);
+  storage.addStorageListener(listeners[key]);
 };
 
 /**
@@ -80,6 +79,6 @@ export const removeCachedDataListener = (
   key: string,
   _updateCallback: (key: string, data: unknown) => void
 ) => {
-  chrome.storage.onChanged.removeListener(listeners[key]);
+  storage.removeStorageListener(listeners[key]);
   delete listeners[key];
 };
