+++ packages/data-model/src/index.ts
@@ -12,4 +12,4 @@ export * from './batch-refresh';
 // Storage
 export * from './storage';
 export * from './storage/storage-types';
-export * from './storage/chrome-storage';
+export * from './storage/memory-storage';
