+++ packages/extension-shared/tsconfig.json
@@ -6,6 +6,9 @@
   "references": [
     {
       "path": "../shared"
+    },
+    {
+      "path": "../data-model"
     }
   ]
 }
